Added check ig logs exist before removing it

+++ main.js
@@ -1,6 +1,8 @@
 const fs = require('fs');
-// let ensure the logs folder is empty
-fs.rmSync("logs", { recursive: true }) 
+// lets ensure the logs folder is empty
+if (fs.existsSync("logs")) {
+  fs.rmSync("logs", { recursive: true }) 
+}
 
 let path = require('path')
 const winston = require(path.join(process.cwd(), 'config/winston.js'));
